optional move validation
The Grid.move method does not perform any validation, to save time
during searching. A separate Grid.validate_move method has been added,
which confirms that a move is valid before performing it, to preserve
the integrity of the game.

The failure state scores for the player AI have been adjusted to 0
instead of -INF, to make sure that the AI chooses something even when
there are no non-terminal moves available.

Assertion statements should be added to the default move methods to
allow for easy debugging of moves.