improved reporting
Additional curses windows were created for displaying running move and
game information. Games are now run in a series, with a median score
and a confidence level calculated. To keep confidence low only a few
games have been played, a configurable amount of noise has been added to
the population sampling.

Confidence is currently calculated as the confidence that the median is
_exactly_ correct (since the possible values are discrete.) An interval
could be calculated as well, but the "interval" should be the size of
the set of possible values.

The current calculated median is within {512,1024} with 95% confidence.